modernize-requirements/evals/fixtures/eval-2-poetry-dockerfile/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.2
click==8.1.8
Flask==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.26.1
packaging==24.2
Pillow==11.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.41
typing-extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
